intplusone Custom Layout Startup File For Blender - Designed Around 2.74
Despite the fact that this was made in Blender 2.74, I suspect that it
should function on earlier versions. If you like this file, or find it
beneficial in your own workflow, please subscribe to intplusone, like us
on Facebook, or send me a personal e-mail at carlrknapp@gmail.com.

I could really use the positive reinforcement!