UI Color Scheme Used at intplusone - Completely Optional
This is a .XML file.

For Windows installation, turn on hidden files and folders in your File
Explorer ( Presuming explorer.exe ) , and navigate to ( in Windows Vista
and and above ) %ROOT%\Users\$USERNAME\Appdata\Roaming\Blender
Foundation\2.74\scripts\presets\interface_theme folder and place the
'intplusone_9_24_15.xml' file into that folder.

To activate, load it in user preserfences - > theme tab in Blender.